fix: resolve 73 compilation errors + Dockerfile fixes for Docker build
Cargo dependencies:
- Add rust_decimal + num-traits to domain and application crates
- Enable uuid v5 feature for deterministic event IDs

Domain/Application/Infrastructure:
- Add missing DomainError variants (InvalidState, InvalidInput, etc.)
- Fix chrono::Datelike/Timelike imports across 4 files
- Fix Currency type mismatch, trait method stubs, type annotations
- Fix Uuid::parse -> parse_str, PrometheusMetrics -> Builder

Dockerfile:
- Upgrade to rust:latest + debian:bookworm-slim runtime
- Fix binary name (server), healthcheck (curl + /api/v1/health)

+++ crates/domain/src/product/entities.rs
@@ -1,5 +1,7 @@
 use chrono::{DateTime, Utc};
+use num_traits::pow::Pow;
 use rust_decimal::Decimal;
+use rust_decimal::prelude::ToPrimitive;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use uuid::Uuid;
@@ -261,7 +263,7 @@ impl InterestRate {
 
     pub fn calculate_compound_monthly(&self, principal: Decimal, months: u32) -> Decimal {
         let monthly_rate = self.annual_rate / Decimal::from(12) / Decimal::from(100);
-        let compound_factor = (Decimal::ONE + monthly_rate).pow(months);
+        let compound_factor = (Decimal::ONE + monthly_rate).pow(months as u64);
         principal * (compound_factor - Decimal::ONE)
     }
 }
